+++ package.json
@@ -27,6 +27,7 @@
     "story:preview": "histoire preview"
   },
   "dependencies": {
+    "@kyvg/vue3-notification": "^2.9.0",
     "@pinia/testing": "^0.0.14",
     "@sigrennesmetropole/cooperation_jn_common_ui": "0.0.12",
     "@tailwindcss/forms": "^0.5.3",
