+++ src/main.ts
@@ -4,6 +4,7 @@ import { createPinia } from 'pinia'
 import App from './App.vue'
 import router from './router'
 import VueApexCharts from 'vue3-apexcharts'
+import Notifications from '@kyvg/vue3-notification'
 
 import './assets/main.css'
 import '@sigrennesmetropole/cooperation_jn_common_ui/dist/style.css'
@@ -20,5 +21,6 @@ const app = createApp(App)
 app.use(createPinia())
 app.use(router)
 app.use(VueApexCharts)
+app.use(Notifications)
 
 app.mount('#app')
